refactoring multi-sig examples
